Move code from lib/ to src/
